Standardize physics on float and centralize constants
- Create Physics.hpp with all GD physics constants as constexpr float
- Change Player struct fields (timeElapsed, acceleration, velocity,
  slopeData.elapsed) from double to float to match GD's internal 32-bit
  precision
- Update roundVel() and setVelocity() to use float
- Update Vehicle.cpp to use float and centralized constants
- Update Slope.cpp/hpp: angle(), expectedY() return float; all internal
  calculations use float; use PHYS_SPEEDS
- Update Block.cpp: change bottom from double to float
- Update pathfinder.cpp to use PHYS_SPEEDS directly
- Pack Player bool flags into bitfields for better cache density

This ensures the simulator's floating-point behavior matches Geometry
Dash exactly, preventing accumulated rounding errors that would cause
the macro to diverge from real gameplay.

+++ gd-sim/include/Slope.hpp
@@ -17,20 +17,19 @@ struct Slope : public Block {
     /// Orientation of the slope relative to the player's gravity.
     int gravOrient(Player const& p) const;
 
-    double angle() const;
+    float angle() const;
     bool isFacingUp() const;
 
     /**
      * The Y position that a player should be snapped to if on the slope.
      * Also used for collision detection
      */
-    virtual double expectedY(Player const& p) const;
+    virtual float expectedY(Player const& p) const;
 };
-
 
 struct SlopeHazard : public Slope {
     using Slope::Slope;
     void collide(Player&) const override;
     bool touching(Player const&) const override;
-    double expectedY(Player const& p) const override;
+    float expectedY(Player const& p) const override;
 };
